add upgrade handler to the pop server
