+++ node/server.go
@@ -547,6 +547,7 @@ func serveHTTP(server *server, l net.Listener, domains []string) {
 	mx.Handle("/rpc", rpcServer)
 	mx.Handle("/", server.localhostHandler())
 	mx.Handle("/p2p/", server.proxyHandler())
+	mx.Handle("/upgrade", http.HandlerFunc(upgradeHandler))
 
 	go func() {
 		s := &http.Server{
